Resolve violations in the search api
